Chance to install dependencies from generator

+++ generators/typescript/index.js
@@ -4,6 +4,17 @@ var chalk = require('chalk');
 var yosay = require('yosay');
 
 module.exports = yeoman.generators.Base.extend({
+
+  constructor: function () {
+    yeoman.generators.Base.apply(this, arguments);
+    var self = this;
+
+    this.installDeps = function() {
+      self.npmInstall();
+      self.spawnCommand('jspm', ['install']);
+    }
+  },
+
   prompting: function () {
     var done = this.async();
     console.info('cssFramework: '+this.option('cssFramework')+ ' vs '+this.options.cssFramework);
@@ -26,18 +37,29 @@ module.exports = yeoman.generators.Base.extend({
       when: function(answers) {
         return answers.typescript;
       }
+    }, {
+      type: 'confirm',
+      name: 'installDeps',
+      message: 'Do you want to automatically install dependencies for npm and jspm?',
+      default: false
     }];
 
     this.prompt(prompts, function (answers) {
       this.props = answers;
       // To access props later use this.props.someOption;
-      // generator returns in typescript was not selected
-      if (answers.typescript == false)
+      /* generator returns in typescript was not selected
+      if (answers.typescript == false) {
+        if (this.props.installDeps) {
+          console.info('Installin from prompting');
+          this.npmInstall();
+        }
         return;
+      }
       else {
         this.conflicter.force = true;
       }
-
+      */
+      this.conflicter.force = answers.typescript;
       this.amd = answers.amd;
       done();
     }.bind(this));
@@ -47,48 +69,54 @@ module.exports = yeoman.generators.Base.extend({
     // copy templates for package.json, build/tasks/build.js
     typescript: function () {
       // this.bulkDirectory('root', '.');
-      this.fs.copyTpl(
-        this.templatePath('root/package.json'),
-        this.destinationPath('package.json'), {
-          githubAccount: this.options.githubAccount,
-          authorName: this.options.authorName,
-          authorEmail: this.options.authorEmail,
-          appDesc: this.options.appDesc,
-          appName: this.options.appName,
-          cssFramework: this.options.cssFramework
-        }
-      );
-      this.fs.copyTpl(
-        this.templatePath('root/tsconfig.json'),
-        this.destinationPath('tsconfig.json'), {
-          amd: this.amd
-        }
-      );
-      this.copy('root/build.js', 'build/tasks/build.js');
+      if (this.props.typescript) {
+        this.fs.copyTpl(
+          this.templatePath('root/package.json'),
+          this.destinationPath('package.json'), {
+            githubAccount: this.options.githubAccount,
+            authorName: this.options.authorName,
+            authorEmail: this.options.authorEmail,
+            appDesc: this.options.appDesc,
+            appName: this.options.appName,
+            cssFramework: this.options.cssFramework
+          }
+        );
+        this.fs.copyTpl(
+          this.templatePath('root/tsconfig.json'),
+          this.destinationPath('tsconfig.json'), {
+            amd: this.amd
+          }
+        );
+        this.copy('root/build.js', 'build/tasks/build.js');
+      }
     },
 
     // See http://yeoman.github.io/generator/actions.html
     typings: function () {
-      this.bulkDirectory('typings', 'typings');
-      this.fs.delete('typings/es6-promise');
-      this.fs.copyTpl(
-        this.templatePath('typings/tsd.d.ts'),
-        this.destinationPath('typings/tsd.d.ts'), {
-          amd: this.amd
-        }
-      );
+      if (this.props.typescript) {
+        this.bulkDirectory('typings', 'typings');
+        this.fs.delete('typings/es6-promise');
+        this.fs.copyTpl(
+          this.templatePath('typings/tsd.d.ts'),
+          this.destinationPath('typings/tsd.d.ts'), {
+            amd: this.amd
+          }
+        );
+      }
     },
 
     // TODO remove src/tsconfig.json
     srcFiles: function () {
-      this.fs.delete('src/*.js');
-      this.bulkDirectory('src', 'src');
-      this.fs.copyTpl(
-        this.templatePath('src/app.ts'),
-        this.destinationPath('src/app.ts'), {
-          cssFramework: this.options.cssFramework
-        }
-      );
+      if (this.props.typescript) {
+        this.fs.delete('src/*.js');
+        this.bulkDirectory('src', 'src');
+        this.fs.copyTpl(
+          this.templatePath('src/app.ts'),
+          this.destinationPath('src/app.ts'), {
+            cssFramework: this.options.cssFramework
+          }
+        );
+      }
     }
 /*
     mapFiles: function () {
@@ -97,5 +125,14 @@ module.exports = yeoman.generators.Base.extend({
       }
     }
 */
+  },
+
+  install: function() {
+    if (this.props.installDeps) {
+      console.info('Installing deps...');
+      // this.npmInstall();
+      // this.spawnCommand
+      this.installDeps();
+    }
   }
 });
